プレイヤーとマンホールフェード

+++ Paper4最新1128/client/source/Animation/AnimationUV.cpp
@@ -124,8 +124,8 @@ void AnimationUV::Update(Vector3 pos, Vector3 angle, float scale)
 		if (nowFlame >= alphaNear)
 		{
 		// 100-100=0  100-50=50   0/50
-		float A = (endFlame - nowFlame);
-		float B =(endFlame - alphaFar);
+		float A = (float)(endFlame - nowFlame);
+		float B =(float)(endFlame - alphaFar);
 		alpha = A / B;
 		alpha = Clamp(alpha, 0.0f, 1.0f);
 
@@ -179,8 +179,8 @@ void AnimationUV::Update(Vector3 pos, Vector3 angle, Vector3 scale)
 		};
 
 		// 0Ç™ìßñæÅ@1Ç™ïsìßñæÇ…Ç∑ÇÈÇÊÇ§Ç…ê›íË
-		float A = (endFlame - nowFlame);
-		float B = (endFlame - alphaNear);
+		float A = (float)(endFlame - nowFlame);
+		float B = (float)(endFlame - alphaNear);
 		alpha = A / B;
 		alpha = Clamp(alpha, 0.0f, 1.0f);
